+++ public/assets/css/custom.css
@@ -31,6 +31,7 @@ width: 100%;
     background: #cccccc;
     border-radius: 50px;
     transition: 300ms linear;
+    margin-left:15%;
   }
   
   .toggle_btn label:before {
