+++ src/Pages/User.js
@@ -5,6 +5,7 @@ import { Link, useNavigate } from "react-router-dom";
 import UserModal from "../Modals/AddModals/UserModal";
 import { userList } from "../Slices/UserSlice";
 import { useDispatch, useSelector } from "react-redux";
+import Loader from "../utils/Loader";
 
 function User() {
   const dispatch = useDispatch();
@@ -43,8 +44,8 @@ function User() {
         <li key={i}>
           <button
             className={`px-3 py-1 rounded-md ${currentPage === i
-                ? "text-white bg-purple-600 border border-r-0 border-purple-600"
-                : "focus:outline-none focus:shadow-outline-purple"
+              ? "text-white bg-purple-600 border border-r-0 border-purple-600"
+              : "focus:outline-none focus:shadow-outline-purple"
               }`}
             onClick={() => handlePageChange(i)}
           >
@@ -79,8 +80,8 @@ function User() {
         <li key={i}>
           <button
             className={`px-3 py-1 rounded-md ${currentPage === i
-                ? "text-white bg-purple-600 border border-r-0 border-purple-600"
-                : "focus:outline-none focus:shadow-outline-purple"
+              ? "text-white bg-purple-600 border border-r-0 border-purple-600"
+              : "focus:outline-none focus:shadow-outline-purple"
               }`}
             onClick={() => handlePageChange(i)}
           >
@@ -134,6 +135,8 @@ function User() {
   }, [trigger, response, header, navigate, dispatch]);
 
   return (
+    <>
+    {loading && <Loader />}
     <div className="flex h-screen bg-gray-50 dark:bg-gray-900">
       <LeftMenuList />
       <div className="flex flex-col flex-1 w-full">
@@ -179,6 +182,7 @@ function User() {
                     <th className="px-4 py-3">User Email Id</th>
                     <th className="px-4 py-3">User phone Number</th>
                     <th className="px-4 py-3">enterprise Name</th>
+                    <th className="px-4 py-3">Action</th>
                   </tr>
                 </thead>
                 <tbody className="bg-white divide-y dark:divide-gray-700 dark:bg-gray-800">
@@ -206,38 +210,22 @@ function User() {
                       {item.type === "System-integrator" && (
                         <td className="px-4 py-3">
                           <div className="flex items-center text-sm">
-                            {/* <!-- Avatar with inset shadow --> */}
-                            <div className="relative hidden w-8 h-8 mr-3 rounded-full md:block">
-                              {/* <img
-                                className="object-cover w-full h-full rounded-full"
-                                src="assets/img/tata-logo (1).png"
-                                alt=""
-                                loading="lazy"
-                              /> */}
-                              <div
-                                className="absolute inset-0 rounded-full shadow-inner"
-                                aria-hidden="true"
-                              ></div>
-                            </div>
+
                             <div>
                               <p className="font-semibold">
                                 ----
-                                {/* {item.systemIntegratorId?.username} */}
                               </p>
                             </div>
                           </div>
                         </td>
-                      )}
+                        // -------------------------------------------------------------------------------------------------------
+                       
+                     
+                     )}
                       {item.type === "EnterpriseUser" && (
                         <td className="px-4 py-3">
                           <div className="flex items-center text-sm">
-                            {/* <!-- Avatar with inset shadow --> */}
-                            <div className="relative hidden w-8 h-8 mr-3 rounded-full md:block">
-                              <div
-                                className="absolute inset-0 rounded-full shadow-inner"
-                                aria-hidden="true"
-                              ></div>
-                            </div>
+
                             <div>
                               <p className="font-semibold">
                                 {item.enterpriseUserId?.username}
@@ -246,6 +234,48 @@ function User() {
                           </div>
                         </td>
                       )}
+                       <td>
+                        <button
+                          className="px-2 py-2 border-2 border-purple-600 text-purple-600 rounded-md"
+                        >
+                          <svg
+                            xmlns="http://www.w3.org/2000/svg"
+                            width="16"
+                            height="16"
+                            viewBox="0 0 24 24"
+                            fill="none"
+                            stroke="currentcolor"
+                            strokeWidth="2"
+                            strokeLinecap="round"
+                            strokeLinejoin="round"
+                          >
+                            <polygon points="16 3 21 8 8 21 3 21 3 16 16 3"></polygon>
+                          </svg>
+                        </button>
+
+                      
+                        <button
+                          className="px-2 py-2 border-2 border-red-600 text-purple-600 rounded-md"
+                          // onClick={() => openDeleteModal(item)}
+                        >
+                          <svg
+                            xmlns="http://www.w3.org/2000/svg"
+                            width="16"
+                            height="16"
+                            viewBox="0 0 24 24"
+                            fill="none"
+                            stroke="red"
+                            strokeWidth="2"
+                            strokeLinecap="round"
+                            strokeLinejoin="round"
+                          >
+                            <polyline points="3 6 5 6 21 6"></polyline>
+                            <path d="M19 6v14a2 2 0 0 1-2 2H7a2 2 0 0 1-2-2V6m3 0V4a2 2 0 0 1 2-2h4a2 2 0 0 1 2 2v2"></path>
+                            <line x1="10" y1="11" x2="10" y2="17"></line>
+                            <line x1="14" y1="11" x2="14" y2="17"></line>
+                          </svg>
+                        </button>
+                      </td>
                     </tr>
                   ))}
                 </tbody>
@@ -311,6 +341,7 @@ function User() {
         </main>
       </div>
     </div>
+    </>
   );
 }
